test(site): share the long-conversation fixture across chat stories
The page and view scrolling stories each defined a near-identical
long-conversation generator. Move a single buildLongConversation(chatId,
count) into storyFixtures.ts and use it from both.
