+++ site/src/pages/AgentsPage/components/ChatConversation/storyFixtures.ts
@@ -18,6 +18,39 @@ export type StoryStreamRenderState = {
 	liveStatus: LiveStatusModel;
 };
 
+/**
+ * Generate a long conversation so the scroll container overflows in
+ * transcript-scrolling stories.
+ */
+export const buildLongConversation = (
+	chatId: string,
+	count: number,
+): TypesGen.ChatMessage[] => {
+	const messages: TypesGen.ChatMessage[] = [];
+	for (let i = 1; i <= count; i++) {
+		const role: TypesGen.ChatMessageRole = i % 2 === 1 ? "user" : "assistant";
+		const turn = Math.ceil(i / 2);
+		messages.push({
+			id: i,
+			chat_id: chatId,
+			created_at: new Date(Date.now() - (count - i) * 60_000).toISOString(),
+			role,
+			content: [
+				{
+					type: "text",
+					text:
+						role === "user"
+							? `Question ${turn}: Can you explain concept ${turn} in detail?`
+							: `Sure! Here is a detailed explanation of concept ${turn}. `.repeat(
+									4,
+								),
+				},
+			],
+		});
+	}
+	return messages;
+};
+
 const DEFAULT_LIVE_STATUS_PARAMS: DeriveLiveStatusParams = {
 	streamState: null,
 	retryState: null,
